Rename BrowserPanelTest -> TaskViewPanelTest
